installing eslint and fixing errors
