+++ src/components/TestResults.jsx
@@ -1,10 +1,8 @@
-import React from 'react';
-import { useState, useEffect } from 'react';
+import { React, useState, useEffect } from 'react';
+import { Link, useParams } from 'react-router-dom';
 import apiClient from '../services/ApiClient';
-import { Link, useParams } from "react-router-dom";
 
-
-const TestResults = () => {
+function TestResults() {
   const testId = useParams().id;
 
   const [results, setResults] = useState([]);
@@ -13,31 +11,30 @@ const TestResults = () => {
     const run = async () => {
       try {
         const test = await apiClient.getTest(testId);
-        console.log("test: ", test);
-        setResults(test)
+        setResults(test);
       } catch (err) {
         console.log(err);
       }
-    }
+    };
     run();
-  }
+  };
 
-  useEffect(getTestHook, [])
+  useEffect(getTestHook, []);
 
   return (
     <div>
       <Link to="/">
-        <button>Home</button>
+        <button type="button">Home</button>
       </Link>
       <h1>TestsResults</h1>
 
-    {results.map((item)=>{
-      return (
-        <div key={item.id}> {JSON.stringify(item)}</div>
-      )
-    })}
+      {results.map((item) => (
+        <div key={item.id}>
+          {JSON.stringify(item)}
+        </div>
+      ))}
     </div>
-  )
+  );
 }
 
 export default TestResults;
