+++ src/index.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import App from './App';
 import {
-  BrowserRouter as Router
-} from "react-router-dom";
+  BrowserRouter as Router,
+} from 'react-router-dom';
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <Router>
     <App />
-  </Router>
+  </Router>,
 );
