+++ src/services/ApiClient.js
@@ -1,16 +1,17 @@
-import axios from "axios";
+/* eslint-disable consistent-return */
+import axios from 'axios';
 
-const URL = "http://testscrud-env.eba-fpb5kcf8.us-east-1.elasticbeanstalk.com";
+const URL = 'http://testscrud-env.eba-fpb5kcf8.us-east-1.elasticbeanstalk.com';
 
 function logError(errorResponse) {
-  const response = errorResponse.response;
+  const { response } = errorResponse;
 
   if (response && response.data && response.data.error) {
     console.error(`HTTP Error: ${response.data.error}`);
   } else {
-    console.error("Error: ", errorResponse);
+    console.error('Error: ', errorResponse);
   }
-};
+}
 
 const apiClient = {
   createTest: async (test) => {
@@ -24,7 +25,7 @@ const apiClient = {
   getTests: async () => {
     try {
       const { data } = await axios.get(`${URL}/api/tests`);
-      return data; 
+      return data;
     } catch (e) {
       logError(e);
     }
@@ -32,7 +33,7 @@ const apiClient = {
   getTest: async (id) => {
     try {
       const { data } = await axios.get(`${URL}/api/tests/${id}`);
-      return data; 
+      return data;
     } catch (e) {
       logError(e);
     }
